pyproject.toml
[tool.poetry]
name = "real-time-chat"
version = "0.1.0"
description = "Real-time chat application using FastAPI and SocketIO"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.78.0"
uvicorn = "^0.17.0"
SQLAlchemy = "^1.4.0"
psycopg2-binary = "^2.9.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-dotenv = "^0.19.0"
pydantic = "^1.8.2"
dependency-injector = "^4.39.0"
fastapi-socketio = "^0.0.10"
redis = "^4.0.0"
aiohttp = "^3.8.0"

[tool.poetry.dev-dependencies]
alembic = "^1.7.5"
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"
httpx = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
